+++ generator_test.go
@@ -9,14 +9,14 @@ import (
 )
 
 // Generates increasing IDs even with decreasing or constant timestamp
-func TestDecreasingOrConstantTimestamp(t *testing.T) {
+func TestDecreasingOrConstantTimestampReset(t *testing.T) {
 	var ts uint64 = 0x0123_4567_89ab
 	var g *Generator = NewGenerator()
 	if g.LastStatus() != GeneratorStatusNotExecuted {
 		t.Fail()
 	}
 
-	prev, _ := g.GenerateCore(ts)
+	prev, _ := g.GenerateOrResetCore(ts, 10_000)
 	if g.LastStatus() != GeneratorStatusNewTimestamp {
 		t.Fail()
 	}
@@ -27,9 +27,9 @@ func TestDecreasingOrConstantTimestamp(t *testing.T) {
 	for i := uint64(0); i < 100_000; i++ {
 		var curr Id
 		if i < 9_998 {
-			curr, _ = g.GenerateCore(ts - i)
+			curr, _ = g.GenerateOrResetCore(ts-i, 10_000)
 		} else {
-			curr, _ = g.GenerateCore(ts - 9_998)
+			curr, _ = g.GenerateOrResetCore(ts-9_998, 10_000)
 		}
 		if g.LastStatus() != GeneratorStatusCounterLoInc &&
 			g.LastStatus() != GeneratorStatusCounterHiInc &&
@@ -46,15 +46,15 @@ func TestDecreasingOrConstantTimestamp(t *testing.T) {
 	}
 }
 
-// Breaks increasing order of IDs if timestamp moves backward a lot
-func TestTimestampRollback(t *testing.T) {
+// Breaks increasing order of IDs if timestamp went backwards by ten seconds
+func TestTimestampRollbackReset(t *testing.T) {
 	var ts uint64 = 0x0123_4567_89ab
 	var g *Generator = NewGenerator()
 	if g.LastStatus() != GeneratorStatusNotExecuted {
 		t.Fail()
 	}
 
-	prev, _ := g.GenerateCore(ts)
+	prev, _ := g.GenerateOrResetCore(ts, 10_000)
 	if g.LastStatus() != GeneratorStatusNewTimestamp {
 		t.Fail()
 	}
@@ -62,7 +62,7 @@ func TestTimestampRollback(t *testing.T) {
 		t.Fail()
 	}
 
-	curr, _ := g.GenerateCore(ts - 10_000)
+	curr, _ := g.GenerateOrResetCore(ts-10_000, 10_000)
 	if g.LastStatus() != GeneratorStatusClockRollback {
 		t.Fail()
 	}
@@ -74,7 +74,7 @@ func TestTimestampRollback(t *testing.T) {
 	}
 
 	prev = curr
-	curr, _ = g.GenerateCore(ts - 10_001)
+	curr, _ = g.GenerateOrResetCore(ts-10_001, 10_000)
 	if g.LastStatus() != GeneratorStatusCounterLoInc &&
 		g.LastStatus() != GeneratorStatusCounterHiInc &&
 		g.LastStatus() != GeneratorStatusTimestampInc {
@@ -86,14 +86,14 @@ func TestTimestampRollback(t *testing.T) {
 }
 
 // Generates increasing IDs even with decreasing or constant timestamp
-func TestDecreasingOrConstantTimestampNoRewind(t *testing.T) {
+func TestDecreasingOrConstantTimestampAbort(t *testing.T) {
 	var ts uint64 = 0x0123_4567_89ab
 	var g *Generator = NewGenerator()
 	if g.LastStatus() != GeneratorStatusNotExecuted {
 		t.Fail()
 	}
 
-	prev, err := g.GenerateCoreNoRewind(ts, 10_000)
+	prev, err := g.GenerateOrAbortCore(ts, 10_000)
 	if err == ErrClockRollback {
 		t.Fail()
 	}
@@ -107,9 +107,9 @@ func TestDecreasingOrConstantTimestampNoRewind(t *testing.T) {
 	for i := uint64(0); i < 100_000; i++ {
 		var curr Id
 		if i < 9_998 {
-			curr, err = g.GenerateCoreNoRewind(ts-i, 10_000)
+			curr, err = g.GenerateOrAbortCore(ts-i, 10_000)
 		} else {
-			curr, err = g.GenerateCoreNoRewind(ts-9_998, 10_000)
+			curr, err = g.GenerateOrAbortCore(ts-9_998, 10_000)
 		}
 		if err == ErrClockRollback {
 			t.Fail()
@@ -129,15 +129,15 @@ func TestDecreasingOrConstantTimestampNoRewind(t *testing.T) {
 	}
 }
 
-// Returns ErrClockRollback if timestamp moves backward a lot
-func TestTimestampRollbackNoRewind(t *testing.T) {
+// Returns error if timestamp went backwards by ten seconds
+func TestTimestampRollbackAbort(t *testing.T) {
 	var ts uint64 = 0x0123_4567_89ab
 	var g *Generator = NewGenerator()
 	if g.LastStatus() != GeneratorStatusNotExecuted {
 		t.Fail()
 	}
 
-	prev, err := g.GenerateCoreNoRewind(ts, 10_000)
+	prev, err := g.GenerateOrAbortCore(ts, 10_000)
 	if err == ErrClockRollback {
 		t.Fail()
 	}
@@ -148,7 +148,7 @@ func TestTimestampRollbackNoRewind(t *testing.T) {
 		t.Fail()
 	}
 
-	_, err = g.GenerateCoreNoRewind(ts-10_000, 10_000)
+	_, err = g.GenerateOrAbortCore(ts-10_000, 10_000)
 	if err != ErrClockRollback {
 		t.Fail()
 	}
@@ -156,7 +156,7 @@ func TestTimestampRollbackNoRewind(t *testing.T) {
 		t.Fail()
 	}
 
-	_, err = g.GenerateCoreNoRewind(ts-10_001, 10_000)
+	_, err = g.GenerateOrAbortCore(ts-10_001, 10_000)
 	if err != ErrClockRollback {
 		t.Fail()
 	}